added policies to editing and deleting tasks

+++ app/Http/Controllers/App/TaskController.php
@@ -17,29 +17,27 @@ class TaskController extends AppController
     public function create(TaskRepository $taskRepository)
     {
         $group_id = auth()->user()->group_id;
-        $group = Group::find($group_id);
-
-        if (Gate::allows('create-task', $group)) {
-            $users = $taskRepository->UsersByGroupId($group_id);
-    
-            return view('task.create', compact('users'));
-        } else {
-            return back();
-        }
+        $users = $taskRepository->UsersByGroupId($group_id);
+        
+        return view('task.create', compact('users'));
     }
-
+    
     public function store(CreateTaskRequest $request)
     {
         $data = $request->validated();
         $group_id = auth()->user()->group_id;
         $group = Group::find($group_id);
 
-        $task = auth()->user()->tasks()->create([
-            'task' => $data['task'],
-            'group_id' => $group_id,
-        ]);
+        if (Gate::allows('create-task', $group)) {
+            $task = auth()->user()->tasks()->create([
+                'task' => $data['task'],
+                'group_id' => $group_id,
+            ]);
 
-        return redirect('/group/home/' . $group_id);
+            return redirect('/group/home/' . $group_id);
+        } else {
+            return back();
+        }
     }
 
     public function show(Task $task)
@@ -56,20 +54,28 @@ class TaskController extends AppController
 
     public function update(UpdateTaskRequest $request, Task $task)
     {
-        $data = $request->validated();
-
-        $task->update([
-            'code' => trim($data['code']),
-            'code_language' => $data['code_language'],
-        ]);
-
-        return redirect('/task/' . $task->id);
+        if (Gate::allows('update-task', $task)) {
+            $data = $request->validated();
+    
+            $task->update([
+                'code' => trim($data['code']),
+                'code_language' => $data['code_language'],
+            ]);
+    
+            return redirect('/task/' . $task->id);
+        } else {
+            return back();
+        }
     }
 
     public function destroy(Task $task)
     {
-        $task->delete();
-
-        return redirect('/group/home/' . $task->group_id);
+        if (Gate::allows('delete-task', $task)) {
+            $task->delete();
+    
+            return redirect('/group/home/' . $task->group_id);
+        } else {
+            return back();
+        }
     }
 }
